requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.1
django-ckeditor==6.7.1
django-cleanup==9.0.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
idna==3.10
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
